[AIP-21] Move contrib.hooks.qubole_hook providers.qubole.hooks.qubole
