+++ airflow/contrib/operators/qubole_operator.py
@@ -20,12 +20,12 @@
 import re
 from typing import FrozenSet, Iterable, Optional
 
-from airflow.contrib.hooks.qubole_hook import (
-    COMMAND_ARGS, HYPHEN_ARGS, POSITIONAL_ARGS, QuboleHook, flatten_list,
-)
 from airflow.hooks.base_hook import BaseHook
 from airflow.models import BaseOperator, BaseOperatorLink
 from airflow.models.taskinstance import TaskInstance
+from airflow.providers.qubole.hooks.qubole import (
+    COMMAND_ARGS, HYPHEN_ARGS, POSITIONAL_ARGS, QuboleHook, flatten_list,
+)
 from airflow.utils.decorators import apply_defaults
