+++ airflow/providers/qubole/hooks/qubole_check.py
@@ -22,8 +22,8 @@ from io import StringIO
 
 from qds_sdk.commands import Command
 
-from airflow.contrib.hooks.qubole_hook import QuboleHook
 from airflow.exceptions import AirflowException
+from airflow.providers.qubole.hooks.qubole import QuboleHook
 from airflow.utils.log.logging_mixin import LoggingMixin
 
 COL_DELIM = '\t'
