+++ tests/contrib/operators/test_qubole_check_operator.py
@@ -23,10 +23,10 @@ from datetime import datetime
 import mock
 from qds_sdk.commands import HiveCommand
 
-from airflow.contrib.hooks.qubole_hook import QuboleHook
 from airflow.contrib.operators.qubole_check_operator import QuboleValueCheckOperator
 from airflow.exceptions import AirflowException
 from airflow.models import DAG
+from airflow.providers.qubole.hooks.qubole import QuboleHook
 from airflow.providers.qubole.hooks.qubole_check import QuboleCheckHook
